Corrigindo erro double na divisao de dois inteiros e incrementando o contador apenas quando instanciamos um novo objeto

+++ src/Pais.java
@@ -85,7 +85,7 @@ public class Pais {
     }
 
     public double calcularDensidade(){
-        return Integer.parseInt(this.populacao) / Integer.parseInt(this.dimensao);
+        return (double) Integer.parseInt(this.populacao) / Integer.parseInt(this.dimensao);
     }
 
     public void paisesVizinhosIguais(Pais b){
